requirements.txt
appdirs==1.4.2
dj-database-url==0.4.1
Django==1.9.7
djangorestframework==3.5.4
gunicorn==19.6.0
packaging==16.8
psycopg2==2.6.1
pyparsing==2.1.10
six==1.10.0
whitenoise==2.0.6
